rename and add some testdata
