+++ src/frontend/src/user/user_manager.rs
